Add nested brackets for ease of environment setup w/ usage comments
The nested brackets group a large portion of the global setup code at the front of the script so that they can be more easily run for measure-specific QA. First chunk includes everything up until the first for-loop, and the second chunk resumes after the for-loop (skipping it) and continues until just before Q1 is reached (after all global variables are defined).

Also set placeholders in script for Q22f objects.